+++ src/mock/destination.js
@@ -1,4 +1,4 @@
-import { getRandomInteger } from '../utils.js';
+import { getRandomInteger } from '../utils/common.js';
 
 const generateDescription = () => {
   const descriptions = [
